requirements.txt
amqp==5.0.6
arrow==1.1.1
asgiref==3.4.1
Babel==2.9.1
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
confusable-homoglyphs==3.2.0
cookiecutter==1.7.3
cryptography==3.4.7
defusedxml==0.7.1
Django<3.2.5
django-allauth==0.45.0
django-anymail==8.4
django-celery-beat==2.2.1
django-cleanup==5.2.0
django-registration==3.2
django-timezone-field==4.2.1
django-widget-tweaks==1.4.8
flower==1.0.0
gunicorn==20.1.0
humanize==3.10.0
idna==3.2
Jinja2==3.0.1
jinja2-time==0.2.0
kombu==5.1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.1
poyo==0.5.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
py-moneyed==1.2
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
django-cors-headers==3.8.0
djangorestframework~=3.12.4